fixed lint error for ChannelDetailScreen.kt
